+++ react/components/pages/Details/DataCard/DisplayData.tsx
@@ -1,12 +1,28 @@
 import React, { FunctionComponent } from 'react'
-import { InjectedIntlProps, injectIntl } from 'react-intl'
+import { InjectedIntlProps, injectIntl, defineMessages } from 'react-intl'
 import { Tag } from 'vtex.styleguide'
 
 import { CSS, BASIC_CARD_WRAPPER } from '../../../../constants'
 import EditButton from '../../../commons/EditButton'
-import FrequencyInfo from '../../../commons/FrequencyInfo'
+import FrequencyInfo from '../../../Frequency/Info'
 import LabeledInfo from '../../../commons/LabeledInfo'
 import { SubscriptionsGroup } from '..'
+
+const messages = defineMessages({
+  nextPurchase: {
+    id: 'store/subscription.nextPurchase',
+    defaultMessage: '',
+  },
+  skip: { id: 'store/subscription.skip.confirm', defaultMessage: '' },
+  estimatedDelivery: {
+    id: 'store/subscription.data.estimatedDelivery',
+    defaultMessage: '',
+  },
+  cardTitle: {
+    id: 'store/subscription.data',
+    defaultMessage: '',
+  },
+})
 
 const DisplayData: FunctionComponent<Props> = ({ group, intl, onOpenEdit }) => {
   let displayDelivery = false
@@ -19,7 +35,7 @@ const DisplayData: FunctionComponent<Props> = ({ group, intl, onOpenEdit }) => {
     <div className={`${BASIC_CARD_WRAPPER} ${CSS.cardHorizontalPadding}`}>
       <div className="flex">
         <div className="db-s di-ns f4 tl c-on-base">
-          {intl.formatMessage({ id: 'subscription.data' })}
+          {intl.formatMessage(messages.cardTitle)}
         </div>
         <div className="ml-auto">
           <EditButton
@@ -29,42 +45,43 @@ const DisplayData: FunctionComponent<Props> = ({ group, intl, onOpenEdit }) => {
           />
         </div>
       </div>
-
       <div className="pt5-s pt5-ns w-100-s mr-auto">
         <FrequencyInfo
           periodicity={group.plan.frequency.periodicity}
           purchaseDay={group.purchaseSettings.purchaseDay}
           interval={group.plan.frequency.interval}
         />
-
         <div className="flex-l">
           <div className="w-50-l pt6">
-            <LabeledInfo labelId="subscription.nextPurchase">
-              <div className="flex flex-row">
+            <LabeledInfo label={intl.formatMessage(messages.nextPurchase)}>
+              <div className="flex">
                 <span className="db fw3 f5-ns f6-s c-on-base">
                   {intl.formatDate(group.nextPurchaseDate)}
                 </span>
                 {group.isSkipped && (
                   <div className="lh-solid mt1 ml3">
                     <Tag type="warning">
-                      {intl.formatMessage({ id: 'subscription.skip.confirm' })}
+                      {intl.formatMessage(messages.skip)}
                     </Tag>
                   </div>
                 )}
               </div>
             </LabeledInfo>
           </div>
-
-          <div className="w-50-l pt6">
-            <LabeledInfo labelId="subscription.data.estimatedDelivery">
-              {displayDelivery &&
-                intl.formatDate(
-                  group.shippingEstimate &&
-                    group.shippingEstimate.estimatedDeliveryDate,
-                  { timeZone: 'UTC' }
+          {displayDelivery && (
+            <div className="w-50-l pt6">
+              <LabeledInfo
+                label={intl.formatMessage(messages.estimatedDelivery)}
+              >
+                {intl.formatDate(
+                  group.shippingEstimate?.estimatedDeliveryDate,
+                  {
+                    timeZone: 'UTC',
+                  }
                 )}
-            </LabeledInfo>
-          </div>
+              </LabeledInfo>
+            </div>
+          )}
         </div>
       </div>
     </div>